requirements.txt
aiosqlite==0.21.0
asyncpg==0.30.0
certifi==2025.10.5
charset-normalizer==3.4.4
cheroot==11.1.2
idna==3.11
iso8601==2.1.0
jaraco.functools==4.3.0
more-itertools==10.8.0
pypika-tortoise==0.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
tortoise-orm==0.25.1
typing_extensions==4.15.0
urllib3==2.5.0
web.py==0.62
xendit-python==7.0.0
